chore: rename supabaseAuthHelpersMock to supabaseSsrMock
