+++ jest.config.mjs
@@ -20,8 +20,7 @@ const config = {
     '~/context/(.*)$': '<rootDir>/src/context/$1',
     '~/utils/(.*)$': '<rootDir>/src/utils/$1',
     '~/ui/(.*)$': '<rootDir>/src/ui/$1',
-    '@supabase/auth-helpers-nextjs':
-      '<rootDir>/src/__tests__/__mocks__/supabaseAuthHelpersMock.ts',
+    '@supabase/ssr': '<rootDir>/src/__tests__/__mocks__/supabaseSsrMock.ts',
   },
   setupFilesAfterEnv: ['<rootDir>/src/__tests__/jest-setup.js'],
   testPathIgnorePatterns: ['<rootDir>/src/e2e', '<rootDir>/src/__tests__'],
